(feat): add deplyoment, Add sourceMap support
