+++ .github/workflows/build_and_deploy.yml
@@ -1,0 +1,32 @@
+name: Deploy to Google Cloud
+
+on:
+  push:
+
+jobs:
+  deploy:
+    runs-on: 'ubuntu-latest'
+    permissions:
+      contents: 'read'
+      id-token: 'write'
+
+    steps:
+      - uses: 'actions/checkout@v3'
+
+      - id: 'auth'
+        uses: 'google-github-actions/auth@v1'
+        with:
+          credentials_json: '${{ secrets.GCP_CREDENTIALS }}'
+
+      - id: 'deploy_function'
+        uses: 'google-github-actions/deploy-cloud-functions@v1'
+        with:
+          name: 'immp-scrape'
+          runtime: 'nodejs18'
+
+      - id: 'deploy_scheduler'
+        run: \
+          gcloud scheduler jobs create http YOUR_SCHEDULER_JOB_NAME \
+          --schedule="0 * * * *" \
+          --http-method=GET \
+          --uri=${{steps.deploy.outputs.url}} \
