+++ package.json
@@ -33,6 +33,7 @@
     "@types/express": "^4.17.15",
     "@types/node": "^18.11.17",
     "gts": "^3.1.1",
+    "source-map-support": "^0.5.21",
     "typescript": "^4.9.4"
   },
   "dependencies": {
