+++ src/functions/scrape/index.ts
@@ -1,13 +1,9 @@
 import { http, Request, Response } from '@google-cloud/functions-framework';
-import WillHabenService, {
-  WillHabenConfig,
-} from '../../services/willHaben.service';
-import {
-  VIENNADISTRICTS,
-  PROPERTYTYPES,
-  FREEAREATYPES,
-} from '../../services/willHaben.types';
+import WillHabenService from '../../services/willHaben.service';
 import DiscordService from '../../services/discord.service';
+
+import 'source-map-support/register';
+import willHabenConfig from './willhaben.config';
 
 const willHabenService = new WillHabenService();
 
@@ -35,42 +31,6 @@ const main = async () => {
     process.env.DISCORD_CHANNEL_ID
   );
 
-  const willHabenConfig: WillHabenConfig = {
-    AREA_ID: [
-      VIENNADISTRICTS.FIRST,
-      VIENNADISTRICTS.SECOND,
-      VIENNADISTRICTS.THIRD,
-      VIENNADISTRICTS.FOURTH,
-      VIENNADISTRICTS.FIFTH,
-      VIENNADISTRICTS.SIXTH,
-      VIENNADISTRICTS.SEVENTH,
-      VIENNADISTRICTS.EIGHTH,
-      VIENNADISTRICTS.NINTH,
-    ],
-    PROPERTY_TYPE: [
-      PROPERTYTYPES.APARTMENT,
-      PROPERTYTYPES.COOPERATIVE_APARTMENT,
-      PROPERTYTYPES.PENTHOUSE,
-      PROPERTYTYPES.MAISONETTE,
-      PROPERTYTYPES.GROUND_FLOOR_APARTMENT,
-      PROPERTYTYPES.TOPFLOOR_APARTMENT,
-      PROPERTYTYPES.LOFT,
-    ],
-    FREE_AREA_TYPE: [
-      FREEAREATYPES.BALCONY,
-      FREEAREATYPES.GARDEN,
-      FREEAREATYPES.LOGGIA,
-      FREEAREATYPES.ROOFTERRACE,
-      FREEAREATYPES.TERRACE,
-      FREEAREATYPES.WINTERGARDEN,
-    ],
-    PRICE_FROM: '700',
-    PRICE_TO: '2500',
-    keyword: 'fernwärme',
-    LIVING_AREA_FROM: '85',
-    TIME_PERIOD: '2',
-  };
-
   let flats = await willHabenService.getFlats(willHabenConfig);
   flats = flats.slice(0, 3);
   await discordService.connect();
@@ -78,11 +38,13 @@ const main = async () => {
     flats.map(async flat => {
       const isMessagePosted = await discordService.isFlatPosted(flats[0].id);
       if (!isMessagePosted) {
-        discordService.publishMessage(discordService.constructEmbed(flat));
+        return discordService.publishMessage(
+          discordService.constructEmbed(flat)
+        );
       }
     })
   );
-  await discordService.disconnect();
+  discordService.disconnect();
 };
 
 main();
